+++ packages/auth/plugin/src/ios/index.ts
@@ -1,3 +1,4 @@
 import { withIosCaptchaUrlTypes } from './urlTypes';
+import { withIosCaptchaOpenUrlFix } from './openUrlFix';
 
-export { withIosCaptchaUrlTypes };
+export { withIosCaptchaUrlTypes, withIosCaptchaOpenUrlFix };