+++ packages/auth/plugin/src/pluginConfig.ts
@@ -1,0 +1,7 @@
+export interface PluginConfigType {
+  ios?: PluginConfigTypeIos;
+}
+
+export interface PluginConfigTypeIos {
+  captchaOpenUrlFix?: undefined | boolean | 'default';
+}